# openapi.yaml
openapi: 3.1.0
info:
  title: GMAO API
  description: API Proxy - Gateway entre frontend et données
  version: 2.2.1
paths:
...
  /stock-items/:
    get:
      tags:
        - stock-items
      summary: List Stock Items
      description: Liste tous les articles en stock avec filtres optionnels
      operationId: list_stock_items_stock_items__get
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Nombre d'éléments à sauter
            default: 0
            title: Skip
          description: Nombre d'éléments à sauter
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            description: Nombre max d'éléments
            default: 100
            title: Limit
          description: Nombre max d'éléments
        - name: family_code
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Filtrer par code famille
            title: Family Code
          description: Filtrer par code famille
        - name: sub_family_code
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Filtrer par code sous-famille
            title: Sub Family Code
          description: Filtrer par code sous-famille
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Recherche par nom ou référence
            title: Search
          description: Recherche par nom ou référence
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItemListItem"
                title: Response List Stock Items Stock Items  Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      tags:
        - stock-items
      summary: Create Stock Item
      description: Crée un nouvel article en stock (legacy ou template-based)
      operationId: create_stock_item_stock_items__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemIn"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /stock-items/{item_id}:
    get:
      tags:
        - stock-items
      summary: Get Stock Item
      description: Récupère un article par ID
      operationId: get_stock_item_stock_items__item_id__get
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            title: Item Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    put:
      tags:
        - stock-items
      summary: Update Stock Item
      description: Met à jour un article existant (respect immutabilité template)
      operationId: update_stock_item_stock_items__item_id__put
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemIn"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - stock-items
      summary: Delete Stock Item
      description: Supprime un article
      operationId: delete_stock_item_stock_items__item_id__delete
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            title: Item Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /suppliers/:
    get:
      tags:
        - suppliers
      summary: List Suppliers
      description: Liste tous les fournisseurs avec filtres optionnels
      operationId: list_suppliers_suppliers__get
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Nombre d'éléments à sauter
            default: 0
            title: Skip
          description: Nombre d'éléments à sauter
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            description: Nombre max d'éléments
            default: 100
            title: Limit
          description: Nombre max d'éléments
        - name: is_active
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: "null"
            description: Filtrer par statut actif
            title: Is Active
          description: Filtrer par statut actif
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Recherche par nom, code ou contact
            title: Search
          description: Recherche par nom, code ou contact
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierListItem"
                title: Response List Suppliers Suppliers  Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      tags:
        - suppliers
      summary: Create Supplier
      description: Crée un nouveau fournisseur
      operationId: create_supplier_suppliers__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierIn"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /suppliers/{supplier_id}:
    get:
      tags:
        - suppliers
      summary: Get Supplier
      description: Récupère un fournisseur par ID
      operationId: get_supplier_suppliers__supplier_id__get
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: string
            title: Supplier Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    put:
      tags:
        - suppliers
      summary: Update Supplier
      description: Met à jour un fournisseur existant
      operationId: update_supplier_suppliers__supplier_id__put
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: string
            title: Supplier Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierIn"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - suppliers
      summary: Delete Supplier
      description: Supprime un fournisseur
      operationId: delete_supplier_suppliers__supplier_id__delete
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: string
            title: Supplier Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /stock-item-suppliers/:
    get:
      tags:
        - stock-item-suppliers
      summary: List Stock Item Suppliers
      description: Liste toutes les références fournisseurs avec filtres optionnels
      operationId: list_stock_item_suppliers_stock_item_suppliers__get
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Nombre d'éléments à sauter
            default: 0
            title: Skip
          description: Nombre d'éléments à sauter
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            description: Nombre max d'éléments
            default: 100
            title: Limit
          description: Nombre max d'éléments
        - name: stock_item_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Filtrer par article
            title: Stock Item Id
          description: Filtrer par article
        - name: supplier_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Filtrer par fournisseur
            title: Supplier Id
          description: Filtrer par fournisseur
        - name: is_preferred
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: "null"
            description: Filtrer par préféré
            title: Is Preferred
          description: Filtrer par préféré
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItemSupplierListItem"
                title: Response List Stock Item Suppliers Stock Item Suppliers  Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      tags:
        - stock-item-suppliers
      summary: Create Stock Item Supplier
      description: Crée une nouvelle référence fournisseur
      operationId: create_stock_item_supplier_stock_item_suppliers__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemSupplierIn"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemSupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /stock-item-suppliers/{ref_id}:
    get:
      tags:
        - stock-item-suppliers
      summary: Get Stock Item Supplier
      description: Récupère une référence fournisseur par ID
      operationId: get_stock_item_supplier_stock_item_suppliers__ref_id__get
      parameters:
        - name: ref_id
          in: path
          required: true
          schema:
            type: string
            title: Ref Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemSupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    put:
      tags:
        - stock-item-suppliers
      summary: Update Stock Item Supplier
      description: Met à jour une référence fournisseur existante
      operationId: update_stock_item_supplier_stock_item_suppliers__ref_id__put
      parameters:
        - name: ref_id
          in: path
          required: true
          schema:
            type: string
            title: Ref Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemSupplierIn"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemSupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - stock-item-suppliers
      summary: Delete Stock Item Supplier
      description: Supprime une référence fournisseur
      operationId: delete_stock_item_supplier_stock_item_suppliers__ref_id__delete
      parameters:
        - name: ref_id
          in: path
          required: true
          schema:
            type: string
            title: Ref Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /stock-item-suppliers/stock-item/{stock_item_id}:
    get:
      tags:
        - stock-item-suppliers
      summary: Get Suppliers By Stock Item
      description: Récupère toutes les références fournisseurs d'un article
      operationId: get_suppliers_by_stock_item_stock_item_suppliers_stock_item__stock_item_id__get
      parameters:
        - name: stock_item_id
          in: path
          required: true
          schema:
            type: string
            title: Stock Item Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItemSupplierOut"
                title:
                  Response Get Suppliers By Stock Item Stock Item Suppliers Stock
                  Item  Stock Item Id  Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /stock-item-suppliers/supplier/{supplier_id}:
    get:
      tags:
        - stock-item-suppliers
      summary: Get Stock Items By Supplier
      description: Récupère toutes les références d'un fournisseur
      operationId: get_stock_items_by_supplier_stock_item_suppliers_supplier__supplier_id__get
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: string
            title: Supplier Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItemSupplierOut"
                title:
                  Response Get Stock Items By Supplier Stock Item Suppliers Supplier  Supplier
                  Id  Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /stock-item-suppliers/{ref_id}/set-preferred:
    post:
      tags:
        - stock-item-suppliers
      summary: Set Preferred Supplier
      description: Définit cette référence comme fournisseur préféré pour l'article
      operationId: set_preferred_supplier_stock_item_suppliers__ref_id__set_preferred_post
      parameters:
        - name: ref_id
          in: path
          required: true
          schema:
            type: string
            title: Ref Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItemSupplierOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StockItemIn:
      properties:
        name:
          type: string
          title: Name
          description: Nom de l'article
        family_code:
          type: string
          maxLength: 20
          title: Family Code
          description: Code famille
        sub_family_code:
          type: string
          maxLength: 20
          title: Sub Family Code
          description: Code sous-famille
        spec:
          anyOf:
            - type: string
              maxLength: 50
            - type: "null"
          title: Spec
          description: Spécification
        dimension:
          anyOf:
            - type: string
            - type: "null"
          title: Dimension
          description: Dimension (auto pour items template)
        quantity:
          anyOf:
            - type: integer
              minimum: 0.0
            - type: "null"
          title: Quantity
          description: Quantité en stock
          default: 0
        unit:
          anyOf:
            - type: string
              maxLength: 50
            - type: "null"
          title: Unit
          description: Unité
        location:
          anyOf:
            - type: string
            - type: "null"
          title: Location
          description: Emplacement
        standars_spec:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Standars Spec
          description: ID spec standard
        manufacturer_item_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Manufacturer Item Id
          description: ID article fabricant
        characteristics:
          anyOf:
            - items:
                type: object
              type: array
            - type: "null"
          title: Characteristics
          description: Caractéristiques (obligatoire si template)
      type: object
      required:
        - name
        - family_code
        - sub_family_code
      title: StockItemIn
      description: Schéma d'entrée pour créer un article en stock
    StockItemListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        ref:
          anyOf:
            - type: string
            - type: "null"
          title: Ref
        family_code:
          type: string
          title: Family Code
        sub_family_code:
          type: string
          title: Sub Family Code
        quantity:
          anyOf:
            - type: integer
            - type: "null"
          title: Quantity
          default: 0
        unit:
          anyOf:
            - type: string
            - type: "null"
          title: Unit
        location:
          anyOf:
            - type: string
            - type: "null"
          title: Location
      type: object
      required:
        - id
        - name
        - family_code
        - sub_family_code
      title: StockItemListItem
      description: Schéma léger pour la liste des articles
    StockItemOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        family_code:
          type: string
          title: Family Code
        sub_family_code:
          type: string
          title: Sub Family Code
        spec:
          anyOf:
            - type: string
            - type: "null"
          title: Spec
        dimension:
          type: string
          title: Dimension
        ref:
          anyOf:
            - type: string
            - type: "null"
          title: Ref
          description: Référence générée automatiquement
        quantity:
          anyOf:
            - type: integer
            - type: "null"
          title: Quantity
          default: 0
        unit:
          anyOf:
            - type: string
            - type: "null"
          title: Unit
        location:
          anyOf:
            - type: string
            - type: "null"
          title: Location
        standars_spec:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Standars Spec
        supplier_refs_count:
          anyOf:
            - type: integer
            - type: "null"
          title: Supplier Refs Count
          description: Nombre de références fournisseurs
          default: 0
        manufacturer_item_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Manufacturer Item Id
      type: object
      required:
        - id
        - name
        - family_code
        - sub_family_code
        - dimension
      title: StockItemOut
      description: Schéma de sortie pour un article en stock
    StockItemSupplierIn:
      properties:
        stock_item_id:
          type: string
          format: uuid
          title: Stock Item Id
          description: ID de l'article en stock
        supplier_id:
          type: string
          format: uuid
          title: Supplier Id
          description: ID du fournisseur
        supplier_ref:
          type: string
          title: Supplier Ref
          description: Référence fournisseur
        unit_price:
          anyOf:
            - type: number
            - type: "null"
          title: Unit Price
          description: Prix unitaire
        min_order_quantity:
          anyOf:
            - type: integer
            - type: "null"
          title: Min Order Quantity
          description: Quantité minimum de commande
          default: 1
        delivery_time_days:
          anyOf:
            - type: integer
            - type: "null"
          title: Delivery Time Days
          description: Délai de livraison en jours
        is_preferred:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Preferred
          description: Fournisseur préféré
          default: false
        manufacturer_item_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Manufacturer Item Id
          description: ID de l'article fabricant
      type: object
      required:
        - stock_item_id
        - supplier_id
        - supplier_ref
      title: StockItemSupplierIn
      description: Schéma d'entrée pour créer/modifier une référence fournisseur
    StockItemSupplierListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        stock_item_id:
          type: string
          format: uuid
          title: Stock Item Id
        supplier_id:
          type: string
          format: uuid
          title: Supplier Id
        supplier_ref:
          type: string
          title: Supplier Ref
        unit_price:
          anyOf:
            - type: number
            - type: "null"
          title: Unit Price
        min_order_quantity:
          anyOf:
            - type: integer
            - type: "null"
          title: Min Order Quantity
          default: 1
        delivery_time_days:
          anyOf:
            - type: integer
            - type: "null"
          title: Delivery Time Days
        is_preferred:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Preferred
          default: false
        stock_item_name:
          anyOf:
            - type: string
            - type: "null"
          title: Stock Item Name
        stock_item_ref:
          anyOf:
            - type: string
            - type: "null"
          title: Stock Item Ref
        supplier_name:
          anyOf:
            - type: string
            - type: "null"
          title: Supplier Name
        supplier_code:
          anyOf:
            - type: string
            - type: "null"
          title: Supplier Code
      type: object
      required:
        - id
        - stock_item_id
        - supplier_id
        - supplier_ref
      title: StockItemSupplierListItem
      description: Schéma léger pour la liste
...
    SupplierIn:
      properties:
        name:
          type: string
          title: Name
          description: Nom du fournisseur
        code:
          anyOf:
            - type: string
            - type: "null"
          title: Code
          description: Code fournisseur
        contact_name:
          anyOf:
            - type: string
            - type: "null"
          title: Contact Name
          description: Nom du contact
        email:
          anyOf:
            - type: string
            - type: "null"
          title: Email
          description: Email
        phone:
          anyOf:
            - type: string
            - type: "null"
          title: Phone
          description: Téléphone
        address:
          anyOf:
            - type: string
            - type: "null"
          title: Address
          description: Adresse
        notes:
          anyOf:
            - type: string
            - type: "null"
          title: Notes
          description: Notes
        is_active:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Active
          description: Fournisseur actif
          default: true
      type: object
      required:
        - name
      title: SupplierIn
      description: Schéma d'entrée pour créer/modifier un fournisseur
    SupplierListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        code:
          anyOf:
            - type: string
            - type: "null"
          title: Code
        contact_name:
          anyOf:
            - type: string
            - type: "null"
          title: Contact Name
        email:
          anyOf:
            - type: string
            - type: "null"
          title: Email
        phone:
          anyOf:
            - type: string
            - type: "null"
          title: Phone
        is_active:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Active
          default: true
      type: object
      required:
        - id
        - name
      title: SupplierListItem
      description: Schéma léger pour la liste
...
    SupplierOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        code:
          anyOf:
            - type: string
            - type: "null"
          title: Code
        contact_name:
          anyOf:
            - type: string
            - type: "null"
          title: Contact Name
        email:
          anyOf:
            - type: string
            - type: "null"
          title: Email
        phone:
          anyOf:
            - type: string
            - type: "null"
          title: Phone
        address:
          anyOf:
            - type: string
            - type: "null"
          title: Address
        notes:
          anyOf:
            - type: string
            - type: "null"
          title: Notes
        is_active:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Active
          default: true
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Updated At
      type: object
      required:
        - id
        - name
      title: SupplierOut
      description: Schéma de sortie pour un fournisseur
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError